[Magnum]: clean-up BaseApplication and move Drawables in separate file
